requirements.txt
# onnxruntime-gpu
insightface
mediapipe
python-socketio[asyncio_client]
